AntWeb is now a source.
There's no bounding box option yet. Mostly because the API doesn't allow for a location search and a species search to be combined. We could do a location search, then internally (in R) filter by supplied species but that's kinda stupid. Will talk to the devs and see how we can improve for the next update.